Move performance measurement and plotting to dedicated source
